Move logging module to adapters package
