+++ app/api/adapters/logging/decodelog.py
