+++ app/api/db/models/__init__.py
@@ -1,8 +1,8 @@
-import api.logging
+import api.adapters.logging
 
 from . import base, user_models
 
-logger = api.logging.get_logger(__name__)
+logger = api.adapters.logging.get_logger(__name__)
 
 # Re-export metadata
 # This is used by tests to create the test database.
