+++ app/api/db/models/user_models.py
@@ -7,10 +7,10 @@ from sqlalchemy import Boolean, Column, Date, Enum, ForeignKey, Text
 from sqlalchemy.dialects import postgresql
 from sqlalchemy.orm import Mapped, relationship
 
-import api.logging
+import api.adapters.logging
 from api.db.models.base import Base, IdMixin, TimestampMixin
 
-logger = api.logging.get_logger(__name__)
+logger = api.adapters.logging.get_logger(__name__)
 
 
 class RoleType(str, enum.Enum):
