+++ app/api/route/healthcheck.py
@@ -6,11 +6,11 @@ from sqlalchemy import text
 from werkzeug.exceptions import ServiceUnavailable
 
 import api.adapters.db.flask_db as flask_db
-import api.logging
+import api.adapters.logging
 from api.route import response
 from api.route.schemas import request_schema
 
-logger = api.logging.get_logger(__name__)
+logger = api.adapters.logging.get_logger(__name__)
 
 
 class HealthcheckSchema(request_schema.OrderedSchema):
