+++ app/api/route/user_route.py
@@ -4,7 +4,7 @@ from apiflask import APIBlueprint
 
 import api.adapters.db as db
 import api.adapters.db.flask_db as flask_db
-import api.logging as logging
+import api.adapters.logging as logging
 import api.services.users as user_service
 from api.auth.api_key_auth import api_key_auth
 from api.db.models.user_models import User
