+++ app/setup.cfg
@@ -19,7 +19,7 @@ ignore =
   # don't use bare except, B001 is more descriptive
   E722
 warn-symbols =
-  logging = Please use api.logging instead of logging.
+  logging = Please use api.adapters.logging instead of logging.
 per-file-ignores =
   # The logging initialization has to import the logging library to override it
   api/logging/*:B1
